requirements.txt
pyinstaller
PyQt5
pyyaml
xmltodict